Refactor Models components to enhance session and dataset management, including the addition of ModelsCenterContent for improved structure and functionality.
