+++ DashAI/front/src/components/models/ModelsContext.jsx
@@ -56,6 +56,7 @@ export function ModelsProvider({ children }) {
   const [selectedModel, setSelectedModel] = useState(null);
   const [configOpen, setConfigOpen] = useState(false);
   const [step, setStep] = useState(0);
+  const [selectedOption, setSelectedOption] = useState(OptionsEnum.NEW);
 
   const selectModel = useCallback((model) => {
     setSelectedModel(model);
