+++ DashAI/front/src/components/threeSectionLayout/SelectOptionMenu.jsx
@@ -6,7 +6,6 @@ import OptionBox from "./OptionBox";
 import { useTranslation } from "react-i18next";
 
 export default function SelectOptionMenu({
-  goToNextStep,
   goToPrevStep = null,
   title,
   subtitle,
@@ -14,6 +13,7 @@ export default function SelectOptionMenu({
   searchBar = false,
   showNoDatasetAlert = false,
   onGoToDatasets = null,
+  goToNextStep = null,
 }) {
   const [search, setSearch] = useState("");
   const filteredOptions = options.filter((option) =>
